+++ core.osgi/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/osgi/NodeResolverService.java
@@ -19,23 +19,29 @@ import java.util.Optional;
 import java.util.UUID;
 
 import com.datastax.oss.driver.api.core.CqlSession;
-import org.osgi.service.component.annotations.*;
 
 import com.ericsson.bss.cassandra.ecchronos.connection.NativeConnectionProvider;
 import com.ericsson.bss.cassandra.ecchronos.core.utils.DriverNode;
 import com.ericsson.bss.cassandra.ecchronos.core.utils.NodeResolver;
 import com.ericsson.bss.cassandra.ecchronos.core.utils.NodeResolverImpl;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.component.annotations.ReferenceCardinality;
+import org.osgi.service.component.annotations.ReferencePolicy;
 
 @Component(service = NodeResolver.class)
 public class NodeResolverService implements NodeResolver
 {
-    @Reference(service = NativeConnectionProvider.class, cardinality = ReferenceCardinality.MANDATORY, policy = ReferencePolicy.STATIC)
+    @Reference(service = NativeConnectionProvider.class,
+            cardinality = ReferenceCardinality.MANDATORY,
+            policy = ReferencePolicy.STATIC)
     private volatile NativeConnectionProvider myNativeConnectionProvider;
 
     private volatile NodeResolver delegateNodeResolver;
 
     @Activate
-    public void activate()
+    public final void activate()
     {
         CqlSession session = myNativeConnectionProvider.getSession();
 
@@ -43,13 +49,13 @@ public class NodeResolverService implements NodeResolver
     }
 
     @Override
-    public Optional<DriverNode> fromIp(InetAddress inetAddress)
+    public final Optional<DriverNode> fromIp(final InetAddress inetAddress)
     {
         return delegateNodeResolver.fromIp(inetAddress);
     }
 
     @Override
-    public Optional<DriverNode> fromUUID(UUID nodeId)
+    public final Optional<DriverNode> fromUUID(final UUID nodeId)
     {
         return delegateNodeResolver.fromUUID(nodeId);
     }