+++ core.osgi/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/osgi/commands/RepairConfigCommand.java
@@ -41,19 +41,19 @@ public class RepairConfigCommand implements Action
     }
 
     @VisibleForTesting
-    RepairConfigCommand(RepairScheduler repairScheduler)
+    RepairConfigCommand(final RepairScheduler repairScheduler)
     {
         myRepairScheduler = repairScheduler;
     }
 
     @Override
-    public Object execute() throws Exception
+    public final Object execute() throws Exception
     {
         printConfig(System.out);
         return null;
     }
 
-    void printConfig(PrintStream out)
+    public final void printConfig(final PrintStream out)
     {
         ShellTable table = createShellTable();
 
@@ -77,7 +77,7 @@ public class RepairConfigCommand implements Action
         return table;
     }
 
-    private List<Object> createRowContent(ScheduledRepairJobView job)
+    private List<Object> createRowContent(final ScheduledRepairJobView job)
     {
         RepairConfiguration config = job.getRepairConfiguration();
         String tableName = job.getTableReference().getKeyspace() + "." + job.getTableReference().getTable();
@@ -91,7 +91,7 @@ public class RepairConfigCommand implements Action
                 );
     }
 
-    private static int sortedByName(ScheduledRepairJobView view1, ScheduledRepairJobView view2)
+    private static int sortedByName(final ScheduledRepairJobView view1, final ScheduledRepairJobView view2)
     {
         TableReference table1 = view1.getTableReference();
         TableReference table2 = view2.getTableReference();